requirements.txt
absl-py==0.10.0
astor==0.8.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dataclasses==0.6
dnspython==2.0.0
Flask==1.1.2
future==0.18.2
gast==0.2.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
implicit==0.4.2
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
llvmlite==0.34.0
Markdown==3.2.2
MarkupSafe==1.1.1
numba==0.51.2
numpy==1.19.0
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylast==4.0.0
pymongo==3.11.0
python-dotenv==0.15.0
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.1
six==1.15.0
spotipy==2.13.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.7.0+cpu
torchaudio==0.7.0
torchvision==0.8.1+cpu
tqdm==4.47.0
typing-extensions==3.7.4.3
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
